News card is now shared
