+++ plugins/visual-qontract/src/components/HomeComponent/FeaturedNews.tsx
@@ -13,6 +13,7 @@ import {
 } from '@material-ui/core';
 import { configApiRef, useApi } from '@backstage/core-plugin-api';
 import AnnouncementIcon from '@material-ui/icons/Announcement';
+import { NewsStoryCard } from '../NewsComponent';
 
 export const FeaturedNews = () => {
   const [news, setNews] = useState<any[]>([]);
@@ -91,25 +92,7 @@ export const FeaturedNews = () => {
     return (
       <React.Fragment>
         {featuredNews.map((story, _index) => (
-          <Card raised className={classes.newsCard} onClick={() => {handleCardClick(story.link.url)}}>
-            <CardActionArea>
-              <CardMedia
-                className={classes.media}
-                image={`${proxyUrl}/${story.image}`}
-                title={story.title}
-              />
-              <CardContent className={classes.cardBody}>
-                <Link
-                  href={story.link.url}
-                  target="_blank"
-                  rel="noopener noreferrer"
-                >
-                  <Typography variant="button">{story.title}</Typography>
-                </Link>
-                <Typography variant="body2">{story.body}</Typography>
-              </CardContent>
-            </CardActionArea>
-          </Card>
+          <NewsStoryCard story={story} />
         ))}
       </React.Fragment>
     );
