+++ plugins/visual-qontract/src/components/NewsComponent/index.ts
@@ -1,1 +1,2 @@
 export  { NewsComponent } from './NewsComponent';
+export  { NewsStoryCard } from './NewsStoryCard';
